+++ tests/support/test_expand_limits.py
@@ -5,8 +5,8 @@ from time_split.settings import auto_expand_limits as settings
 from time_split.support import expand_limits
 
 
-def _run(expected, expand_limits, limits):
-    actual = expand_limits(limits, expand_limits=expand_limits)
+def _run(expected, spec, limits):
+    actual = expand_limits(limits, spec=spec)
     assert actual == tuple(Timestamp(s) for s in expected.split(","))
 
 
@@ -33,27 +33,27 @@ def test_auto(actual, expected):
 
 
 @pytest.mark.parametrize(
-    "expand_limits, expected",
+    "spec, expected",
     [
         ("d", "2019-05-11, 2019-05-12"),
         ("h", "2019-05-11, 2019-05-11 23"),
     ],
 )
-def test_basic(expand_limits, expected):
+def test_basic(spec, expected):
     limits = Timestamp("2019-05-11 00:15"), Timestamp("2019-05-11 22:35")
-    _run(expected, expand_limits, limits)
+    _run(expected, spec, limits)
 
 
 @pytest.mark.parametrize(
-    "expand_limits, expected",
+    "spec, expected",
     [
         ("d<1h", "2019-05-11, 2019-05-12"),
         ("d < 10 minutes", "2019-05-11 00:15, 2019-05-11 23:35"),
     ],
 )
-def test_tolerance(expand_limits, expected):
+def test_tolerance(spec, expected):
     limits = Timestamp("2019-05-11 00:15"), Timestamp("2019-05-11 23:35")
-    _run(expected, expand_limits, limits)
+    _run(expected, spec, limits)
 
 
 @pytest.mark.parametrize(
@@ -82,14 +82,14 @@ class TestErrors:
     )
     def test_bad_limits(self, start, stop):
         limits = pd.Timestamp(start), pd.Timestamp(stop)
-        expand_limits = ("this", "shouldn't", "matter")
+        spec = ("this", "shouldn't", "matter")
         with pytest.raises(ValueError):
-            expand_limits(limits, expand_limits=expand_limits)
+            expand_limits(limits, spec=spec)
 
     @pytest.mark.parametrize("round_to", ["1d", "2d", "-d"])
     def test_invalid_round_to_frequency(self, round_to):
         with pytest.raises(ValueError, match="not a valid frequency"):
-            expand_limits(self.limits, expand_limits=("2d", round_to, "3h"))
+            expand_limits(self.limits, spec=("2d", round_to, "3h"))
 
     @pytest.mark.parametrize(
         "start_at, tolerance",
@@ -100,7 +100,7 @@ class TestErrors:
     )
     def test_negative(self, start_at, tolerance):
         with pytest.raises(ValueError, match="non-negative"):
-            expand_limits(self.limits, expand_limits=(start_at, "d", tolerance))
+            expand_limits(self.limits, spec=(start_at, "d", tolerance))
 
     @pytest.mark.parametrize(
         "order, start_at, round_to, tolerance",
@@ -112,7 +112,7 @@ class TestErrors:
     def test_bad_inequality(self, start_at, round_to, tolerance, order):
         left, right = order.split(",")
         with pytest.raises(ValueError, match=f"{left}=.* < {right}="):
-            expand_limits(self.limits, expand_limits=(start_at, round_to, tolerance))
+            expand_limits(self.limits, spec=(start_at, round_to, tolerance))
 
     def test_bad_level(self):
         with pytest.raises(AttributeError, match="Bad level='unknown'"):
@@ -121,9 +121,9 @@ class TestErrors:
     def test_bad_hour_setting(self, monkeypatch):
         monkeypatch.setattr(settings, "hour", ("2 days", "day", "15 min"))
         with pytest.raises(ValueError, match="Invalid settings: auto_expand_limits.hour="):
-            expand_limits(self.limits, expand_limits=True)
+            expand_limits(self.limits, spec=True)
 
     def test_bad_day_setting(self, monkeypatch):
         monkeypatch.setattr(settings, "day", ("2 hour", "hour", "1 hour"))
         with pytest.raises(ValueError, match="Invalid settings: auto_expand_limits.day="):
-            expand_limits(self.limits, expand_limits=True)
+            expand_limits(self.limits, spec=True)